+: hud hp *: some combat UX

+++ Assets/Scripts/Character/BodyTarget.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 public class BodyTarget : WeaponTarget {
 
-    public event Action<int> OnDamageTaken = (damage) => { };
+    public event Action<float> OnDamageTaken = (damage) => { };
 
     public Material damageTakingMaterial;
 
@@ -17,11 +17,11 @@ public class BodyTarget : WeaponTarget {
 
     public bool enableDebugs = false;
 
-    protected override void OnHitReceived(int weaponDamage) {
+    protected override void OnHitReceived(float weaponDamage) {
         TakeDamage(weaponDamage);
     }
 
-    private void TakeDamage(int damage) {
+    private void TakeDamage(float damage) {
         if(enableDebugs)
             Debug.Log("damage taken " + damage, gameObject);
         OnDamageTaken(damage);
